done lab

+++ app/controllers/toys_controller.rb
@@ -1,5 +1,5 @@
 class ToysController < ApplicationController
-  wrap_parameters format: []
+  rescue_from ActiveRecord::RecordNotFound, with: :render_not_found
 
   def index
     toys = Toy.all
@@ -7,19 +7,28 @@ class ToysController < ApplicationController
   end
 
   def create
-    toy = Toys.create(toy_params)
+    toy = Toy.create(toy_params)
     render json: toy, status: :created
   end
 
   def update
     toy = Toy.find_by(id: params[:id])
-    toy.update(toy_params)
+    if toy
+      toy.update(toy_params)
+      render json: toy
+    else
+      render_not_found
+    end
   end
 
   def destroy
     toy = Toy.find_by(id: params[:id])
-    toy.destroy
-    head :no_content
+    if toy
+      toy.destroy
+      head :no_content
+    else
+      render_not_found
+    end
   end
 
   private
@@ -28,4 +37,7 @@ class ToysController < ApplicationController
     params.permit(:name, :image, :likes)
   end
 
+  def render_not_found
+    render json: { error: "Toy not found" }, status: :not_found
+  end
 end
